backend-brain/requirements.txt
fastapi
uvicorn
pydantic
psycopg2-binary
paho-mqtt
torch
numpy
pandas
scikit-learn